refactor: separando Entidades
