+++ Entidades/.gitignore
@@ -1,0 +1,4 @@
+/Cliente.class
+/FormaDePagamento.class
+/Pedido.class
+/Pizza.class
